xz: up to 5.2.2 version.

+++ tools/xz/src/liblzma/lz/lz_encoder.c
@@ -139,7 +139,7 @@ fill_window(lzma_coder *coder, const lzma_allocator *allocator,
 			&& coder->mf.read_pos < coder->mf.read_limit) {
 		// Match finder may update coder->pending and expects it to
 		// start from zero, so use a temporary variable.
-		const size_t pending = coder->mf.pending;
+		const uint32_t pending = coder->mf.pending;
 		coder->mf.pending = 0;
 
 		// Rewind read_pos so that the match finder can hash
@@ -432,7 +432,7 @@ lz_encoder_init(lzma_mf *mf, const lzma_allocator *allocator,
 	} else {
 /*
 		for (uint32_t i = 0; i < mf->hash_count; ++i)
-		mf->hash[i] = EMPTY_HASH_VALUE;
+			mf->hash[i] = EMPTY_HASH_VALUE;
 */
 		memzero(mf->hash, mf->hash_count * sizeof(uint32_t));
 	}
